Added tests to verify getting data for Current Weather works fine. Refactored code accordingly. Removed unnecessary classes from the web tier.

+++ YieldWeather.Domain/CurrentWeatherContract.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
 
 namespace YieldWeather.Domain
 {
@@ -14,26 +11,26 @@ namespace YieldWeather.Domain
         /// <summary>
         /// Minimum Temperature
         /// </summary>
-        public Decimal MinTemp { get; set; }
+        public double MinTemp { get; set; }
         /// <summary>
         /// Maximum Temperature
         /// </summary>
-        public Decimal MaxTemp { get; set; }
+        public double MaxTemp { get; set; }
 
         /// <summary>
         /// Air Pressure
         /// </summary>
-        public Decimal AirPressure { get; set; }
+        public double AirPressure { get; set; }
 
         /// <summary>
         /// Humidity
         /// </summary>
-        public Int32 Humidity { get; set; }
+        public int Humidity { get; set; }
 
 
         /// <summary>
         /// Rainfall
         /// </summary>
-        public Int32 Rainfall { get; set; }
+        public double Rainfall { get; set; }
     }
 }
